+++ packages/create-app.template.tutorial-todo-app.beta/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-app.template.tutorial-todo-app.beta
 
+## 2.1.0-beta.15
+
+## 2.1.0-beta.14
+
 ## 2.1.0-beta.13
 
 ## 2.1.0-beta.12