+++ packages/example-generator/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @osdk/example-generator
 
+## 0.9.0-beta.14
+
+### Patch Changes
+
+- Updated dependencies [776ae5d]
+  - @osdk/create-widget@0.2.0-beta.0
+  - @osdk/create-app@2.1.0-beta.15
+
+## 0.9.0-beta.13
+
+### Patch Changes
+
+- @osdk/create-app@2.1.0-beta.14
+
 ## 0.9.0-beta.12
 
 ### Patch Changes