+++ packages/generator-converters/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @osdk/generator-converters
 
+## 2.1.0-beta.15
+
+### Patch Changes
+
+- @osdk/api@2.1.0-beta.15
+
+## 2.1.0-beta.14
+
+### Minor Changes
+
+- 0a7c747: Add support for timeseries sensors.
+- 91cfa3a: Fixing url parsing for client.
+- 6b6c8a1: Updates gateway endpoints and changes fetchContents for attachments to return a response
+
+### Patch Changes
+
+- Updated dependencies [0c56d21]
+- Updated dependencies [6b6c8a1]
+  - @osdk/api@2.1.0-beta.14
+
 ## 2.1.0-beta.13
 
 ### Patch Changes